replace old builder with new one
