+++ lib/tzdata/far_future_dynamic_periods.ex
@@ -52,11 +52,11 @@ defmodule Tzdata.FarFutureDynamicPeriods do
       false -> year
     end
 
-    from_standard_time = PeriodBuilder.standard_time_from_utc(from, utc_off)
-    from_wall_time = PeriodBuilder.wall_time_from_utc(from, utc_off, std_off)
+    from_standard_time = standard_time_from_utc(from, utc_off)
+    from_wall_time = wall_time_from_utc(from, utc_off, std_off)
     until_utc = PeriodBuilder.datetime_to_utc(Util.time_for_rule(end_rule, until_time_year), utc_off, std_off)
-    until_standard_time = PeriodBuilder.standard_time_from_utc(until_utc, utc_off)
-    until_wall_time = PeriodBuilder.wall_time_from_utc(until_utc, utc_off, std_off)
+    until_standard_time = standard_time_from_utc(until_utc, utc_off)
+    until_wall_time = wall_time_from_utc(until_utc, utc_off, std_off)
 
     period = %{
       std_off: std_off,
@@ -90,11 +90,11 @@ defmodule Tzdata.FarFutureDynamicPeriods do
     from = PeriodBuilder.datetime_to_utc(Util.time_for_rule(rule_end_of_year, year-1), utc_off, std_off_before)
     letter = rule_end_of_year.letter
 
-    from_standard_time = PeriodBuilder.standard_time_from_utc(from, utc_off)
-    from_wall_time = PeriodBuilder.wall_time_from_utc(from, utc_off, std_off)
+    from_standard_time = standard_time_from_utc(from, utc_off)
+    from_wall_time = wall_time_from_utc(from, utc_off, std_off)
     until_utc = PeriodBuilder.datetime_to_utc(Util.time_for_rule(rule_beginning_of_year, year), utc_off, std_off)
-    until_standard_time = PeriodBuilder.standard_time_from_utc(until_utc, utc_off)
-    until_wall_time = PeriodBuilder.wall_time_from_utc(until_utc, utc_off, std_off)
+    until_standard_time = standard_time_from_utc(until_utc, utc_off)
+    until_wall_time = wall_time_from_utc(until_utc, utc_off, std_off)
 
     period = %{
       std_off: std_off,
@@ -117,5 +117,15 @@ defmodule Tzdata.FarFutureDynamicPeriods do
     rules
     |> Util.rules_for_year(year)
     |> Enum.sort(&(&1.in < &2.in))
+  end
+
+  def standard_time_from_utc(atom, _) when is_atom(atom), do: atom
+  def standard_time_from_utc(utc_time, utc_off) do
+    utc_time + utc_off
+  end
+
+  def wall_time_from_utc(atom, _, _) when is_atom(atom), do: atom
+  def wall_time_from_utc(utc_time, utc_offset, standard_offset) do
+    utc_time + utc_offset + standard_offset
   end
 end
