+++ test/tz_period_builder_test.exs
@@ -1,94 +1,408 @@
-defmodule TzPeriodBuilderTest do
+defmodule Tzdata.PeriodBuilderTest do
   use ExUnit.Case, async: true
-  alias Tzdata.PeriodBuilder, as: TzPeriodBuilder
+  import Tzdata.PeriodBuilder
 
   @source_data_dir "test/tzdata_fixtures/source_data/"
-  test "get periods for a zone" do
-    {:ok, map} = Tzdata.BasicDataMap.from_files_in_dir(@source_data_dir)
-    prds = TzPeriodBuilder.calc_periods(map, "Europe/Copenhagen")
-    # First period. Based on Copenhagen 1st zone line
-    assert hd(prds) == %{std_off: 0,
-       from: %{utc: :min, wall: :min, standard: :min},
-       until: %{standard: 59642697600, wall: 59642697600, utc: 59642694580},
-       utc_off: 3020, zone_abbr: "LMT"}
-    # Second period. Based on Copenhagen 2nd zone line
-    assert prds |> Enum.at(1) == %{std_off: 0,
-       from: %{standard: 59642697600, wall: 59642697600, utc: 59642694580},
-       until: %{standard: 59768928000, wall: 59768928000, utc: 59768924980},
-       utc_off: 3020, zone_abbr: "CMT"}
-    # Third period. Based on Copenhagen 3rd zone line
-    assert prds  |> Enum.at(2) == %{std_off: 0,
-       from: %{standard: 59768928580, wall: 59768928580, utc: 59768924980},
-       until: %{standard: 60474726000, wall: 60474726000, utc: 60474722400},
-       utc_off: 3600, zone_abbr: "CET"}
-    # Fourth period. Based on Copenhagen 3rd zone line
-    # Summer time 1916
-    assert prds |> Enum.at(3) == %{std_off: 3600,
-       from: %{standard: 60474726000, wall: 60474729600, utc: 60474722400},
-       until: %{standard: 60486732000, wall: 60486735600, utc: 60486728400},
-       utc_off: 3600, zone_abbr: "CEST"}
-    # Fifth period. Based on Copenhagen 3rd zone line. And rules line 2&3
-    # Period without summer time. From Sep 30 1916 to May 15 1940
-    assert prds |> Enum.at(4) == %{std_off: 0,
-       from: %{standard: 60486732000, wall: 60486732000, utc: 60486728400},
-       until: %{standard: 61232112000, wall: 61232112000, utc: 61232108400},
-       utc_off: 3600, zone_abbr: "CET"}
 
-    # Sixth period. Summer time from 1940 to 1942!
-    assert prds |> Enum.at(5) == %{std_off: 3600,
-       from: %{standard: 61232112000, wall: 61232115600, utc: 61232108400},
-       until: %{standard: 61309965600, wall: 61309969200, utc: 61309962000},
-       utc_off: 3600, zone_abbr: "CEST"}
-
-    # Seventh period. Winter time from 1942 to 1943
-    assert prds |> Enum.at(6) == %{std_off: 0,
-       from: %{standard: 61309965600, wall: 61309965600, utc: 61309962000},
-       until: %{standard: 61322666400, wall: 61322666400, utc: 61322662800},
-       utc_off: 3600, zone_abbr: "CET"}
-
-     # Twentyfirst period. Wintertime from 1948 to 1980
-    assert prds |> Enum.at(20) == %{std_off: 0,
-       from: %{standard: 61491924000, wall: 61491924000, utc: 61491920400},
-       until: %{standard: 62482752000, wall: 62482752000, utc: 62482748400},
-       utc_off: 3600, zone_abbr: "CET"}
-
-     # Twentyseventh period. Summertime in 1982
-    assert prds |> Enum.at(26) == %{std_off: 3600,
-       from: %{standard: 62553348000, wall: 62553351600, utc: 62553344400},
-       until: %{standard: 62569072800, wall: 62569076400, utc: 62569069200},
-       utc_off: 3600, zone_abbr: "CEST"}
-
-    prds = TzPeriodBuilder.calc_periods(map, "Europe/Paris")
-    assert hd(prds) == %{std_off: 0,
-      from: %{utc: :min, wall: :min, standard: :min},
-      until: %{standard: 59680540860, wall: 59680540860, utc: 59680540299},
-      utc_off: 561, zone_abbr: "LMT"}
+  @doc "Convert an ISO time to Gregorian seconds"
+  defmacro sigil_G({:<<>>, _, [string]}, []) do
+    string
+    |> from_iso8601
+    |> :calendar.datetime_to_gregorian_seconds()
+    |> Macro.escape()
   end
 
-  test "get periods for a zone with just one zone line" do
-    {:ok, map} = Tzdata.BasicDataMap.from_files_in_dir(@source_data_dir)
-    prds = TzPeriodBuilder.calc_periods(map, "Etc/UTC")
-    assert hd(prds) == %{std_off: 0, from: %{utc: :min, standard: :min, wall: :min}, until: %{utc: :max, standard: :max, wall: :max}, utc_off: 0, zone_abbr: "UTC"}
-    prds = TzPeriodBuilder.calc_periods(map, "Etc/GMT-10")
-    assert hd(prds) == %{std_off: 0, from: %{utc: :min, standard: :min, wall: :min}, until: %{utc: :max, standard: :max, wall: :max}, utc_off: 36000, zone_abbr: "GMT-10"}
+  defp from_iso8601(
+         <<year::4-bytes, ?-, month::2-bytes, ?-, day::2-bytes, ?T, hour::2-bytes, ?:,
+           min::2-bytes, ?:, sec::2-bytes>>
+       ) do
+    {year, ""} = Integer.parse(year)
+    {month, ""} = Integer.parse(month)
+    {day, ""} = Integer.parse(day)
+    {hour, ""} = Integer.parse(hour)
+    {min, ""} = Integer.parse(min)
+    {sec, ""} = Integer.parse(sec)
+    {{year, month, day}, {hour, min, sec}}
   end
 
-  test "calculate periods for zone where last line has no rules" do
+  setup_all do
     {:ok, map} = Tzdata.BasicDataMap.from_files_in_dir(@source_data_dir)
-    periods = TzPeriodBuilder.calc_periods(map, "Africa/Abidjan")
-    assert length(periods) == 2
+    {:ok, %{map: map}}
   end
 
-  test "calculate periods for zone where last line has rules, but the rules do not continue forever" do
-    {:ok, map} = Tzdata.BasicDataMap.from_files_in_dir(@source_data_dir)
-    periods = TzPeriodBuilder.calc_periods(map, "Asia/Tokyo")
-    assert periods|>Enum.at(11) == %{from: %{standard: 61589206800, utc: 61589174400, wall: 61589206800}, std_off: 0, until: %{standard: :max, utc: :max, wall: :max}, utc_off: 32400, zone_abbr: "JST"}
+  test "can calculate for zones with one line", %{map: map} do
+    periods = calc_periods(map, "Etc/UTC")
+
+    assert periods == [
+             %{
+               std_off: 0,
+               utc_off: 0,
+               zone_abbr: "UTC",
+               from: %{utc: :min, standard: :min, wall: :min},
+               until: %{utc: :max, standard: :max, wall: :max}
+             }
+           ]
+
+    periods = calc_periods(map, "Etc/GMT-10")
+
+    assert periods == [
+             %{
+               std_off: 0,
+               utc_off: 36000,
+               zone_abbr: "GMT-10",
+               from: %{utc: :min, standard: :min, wall: :min},
+               until: %{utc: :max, standard: :max, wall: :max}
+             }
+           ]
   end
 
-  test "calculate periods for zone where in a zone line there is a rule which is an amount of time" do
-    {:ok, map} = Tzdata.BasicDataMap.from_files_in_dir(@source_data_dir)
-    periods = TzPeriodBuilder.calc_periods(map, "Africa/Ceuta")
-    assert periods |> Enum.at(5) == %{from: %{standard: 60724767600, utc: 60724767600, wall: 60724771200}, std_off: 3600,
-             until: %{standard: 60739542000, utc: 60739542000, wall: 60739545600}, utc_off: 0, zone_abbr: "WEST"}
+  test "can calculate for zones with multiple lines", %{map: map} do
+    periods = calc_periods(map, "Africa/Abidjan")
+    assert [first, second] = periods
+
+    assert first == %{
+             std_off: 0,
+             utc_off: -968,
+             zone_abbr: "LMT",
+             from: %{utc: :min, standard: :min, wall: :min},
+             until: %{
+               utc: ~G[1912-01-01T00:16:08],
+               standard: ~G[1912-01-01T00:00:00],
+               wall: ~G[1912-01-01T00:00:00]
+             }
+           }
+
+    assert second == %{
+             std_off: 0,
+             utc_off: 0,
+             zone_abbr: "GMT",
+             from: %{
+               utc: ~G[1912-01-01T00:16:08],
+               standard: ~G[1912-01-01T00:16:08],
+               wall: ~G[1912-01-01T00:16:08]
+             },
+             until: %{utc: :max, standard: :max, wall: :max}
+           }
+  end
+
+  test "can calculate for zones with static offset rules", %{map: map} do
+    periods = calc_periods(map, "Atlantic/Cape_Verde")
+    assert [_lmt, _cvt_1, cvst, cvt_2, _cvt_3] = periods
+
+    assert cvst == %{
+             std_off: 3600,
+             utc_off: -7200,
+             zone_abbr: "CVST",
+             from: %{
+               utc: ~G[1942-09-01T02:00:00],
+               standard: ~G[1942-09-01T00:00:00],
+               wall: ~G[1942-09-01T01:00:00]
+             },
+             until: %{
+               utc: ~G[1945-10-15T01:00:00],
+               standard: ~G[1945-10-14T23:00:00],
+               wall: ~G[1945-10-15T00:00:00]
+             }
+           }
+
+    assert cvt_2 == %{
+             std_off: 0,
+             utc_off: -7200,
+             zone_abbr: "CVT",
+             from: %{
+               utc: ~G[1945-10-15T01:00:00],
+               standard: ~G[1945-10-14T23:00:00],
+               wall: ~G[1945-10-14T23:00:00]
+             },
+             until: %{
+               utc: ~G[1975-11-25T04:00:00],
+               standard: ~G[1975-11-25T02:00:00],
+               wall: ~G[1975-11-25T02:00:00]
+             }
+           }
+  end
+
+  test "can calculate simple DST rules going into the future", %{map: map} do
+    periods = calc_periods(map, "Antarctica/Troll")
+    [_zzz, std_1, dst_1, std_2 | _] = periods
+
+    assert std_1 == %{
+             std_off: 0,
+             utc_off: 0,
+             zone_abbr: "UTC",
+             from: %{
+               utc: ~G[2005-02-12T00:00:00],
+               standard: ~G[2005-02-12T00:00:00],
+               wall: ~G[2005-02-12T00:00:00]
+             },
+             until: %{
+               utc: ~G[2005-03-27T01:00:00],
+               standard: ~G[2005-03-27T01:00:00],
+               wall: ~G[2005-03-27T01:00:00]
+             }
+           }
+
+    assert dst_1 == %{
+             std_off: 7200,
+             utc_off: 0,
+             zone_abbr: "CEST",
+             from: %{
+               utc: ~G[2005-03-27T01:00:00],
+               standard: ~G[2005-03-27T01:00:00],
+               wall: ~G[2005-03-27T03:00:00]
+             },
+             until: %{
+               utc: ~G[2005-10-30T01:00:00],
+               standard: ~G[2005-10-30T01:00:00],
+               wall: ~G[2005-10-30T03:00:00]
+             }
+           }
+
+    assert std_2 == %{
+             std_off: 0,
+             utc_off: 0,
+             zone_abbr: "UTC",
+             from: %{
+               utc: ~G[2005-10-30T01:00:00],
+               standard: ~G[2005-10-30T01:00:00],
+               wall: ~G[2005-10-30T01:00:00]
+             },
+             until: %{
+               utc: ~G[2006-03-26T01:00:00],
+               standard: ~G[2006-03-26T01:00:00],
+               wall: ~G[2006-03-26T01:00:00]
+             }
+           }
+
+    # 40 years, 2 DST transitions per year
+    assert length(periods) > 80
+  end
+
+  test "calculate periods for zone where last line has no rules", %{map: map} do
+    assert [_, _] = calc_periods(map, "Africa/Abidjan")
+  end
+
+  test "calculates correct abbreviation when changing no rule", %{map: map} do
+    periods = calc_periods(map, "America/Chihuahua")
+    # changed from static CST to CST/CDT at the start of 1996
+    assert Enum.at(periods, 6) ==
+             %{
+               std_off: 0,
+               utc_off: -21600,
+               zone_abbr: "CST",
+               from: %{
+                 wall: ~G[1932-04-01T01:00:00],
+                 standard: ~G[1932-04-01T01:00:00],
+                 utc: ~G[1932-04-01T07:00:00]
+               },
+               until: %{
+                 wall: ~G[1996-04-07T02:00:00],
+                 standard: ~G[1996-04-07T02:00:00],
+                 utc: ~G[1996-04-07T08:00:00]
+               }
+             }
+
+    assert Enum.at(periods, 7) ==
+             %{
+               std_off: 3600,
+               utc_off: -21600,
+               zone_abbr: "CDT",
+               from: %{
+                 wall: ~G[1996-04-07T03:00:00],
+                 standard: ~G[1996-04-07T02:00:00],
+                 utc: ~G[1996-04-07T08:00:00]
+               },
+               until: %{
+                 wall: ~G[1996-10-27T02:00:00],
+                 standard: ~G[1996-10-27T01:00:00],
+                 utc: ~G[1996-10-27T07:00:00]
+               }
+             }
+
+    assert Enum.at(periods, 8) ==
+             %{
+               std_off: 0,
+               utc_off: -21600,
+               zone_abbr: "CST",
+               from: %{
+                 wall: ~G[1996-10-27T01:00:00],
+                 standard: ~G[1996-10-27T01:00:00],
+                 utc: ~G[1996-10-27T07:00:00]
+               },
+               until: %{
+                 wall: ~G[1997-04-06T02:00:00],
+                 standard: ~G[1997-04-06T02:00:00],
+                 utc: ~G[1997-04-06T08:00:00]
+               }
+             }
+  end
+
+  test "can handle DST rules that aren't active for a year", %{map: map} do
+    periods = calc_periods(map, "America/Regina")
+    # -7:00 (no offset) in Sep 1905, since there was no Canadian DST utnil 1918
+    [_lmt, mst_1905_to_1918, mdt_1918, mst_1918_to_1930 | _] = periods
+
+    assert mst_1905_to_1918 == %{
+             std_off: 0,
+             utc_off: -25200,
+             zone_abbr: "MST",
+             from: %{
+               wall: ~G[1905-08-31T23:58:36],
+               standard: ~G[1905-08-31T23:58:36],
+               utc: ~G[1905-09-01T06:58:36]
+             },
+             until: %{
+               wall: ~G[1918-04-14T02:00:00],
+               standard: ~G[1918-04-14T02:00:00],
+               utc: ~G[1918-04-14T09:00:00]
+             }
+           }
+
+    assert mdt_1918 == %{
+             std_off: 3600,
+             utc_off: -25200,
+             zone_abbr: "MDT",
+             from: %{
+               wall: ~G[1918-04-14T03:00:00],
+               standard: ~G[1918-04-14T02:00:00],
+               utc: ~G[1918-04-14T09:00:00]
+             },
+             until: %{
+               wall: ~G[1918-10-27T02:00:00],
+               standard: ~G[1918-10-27T01:00:00],
+               utc: ~G[1918-10-27T08:00:00]
+             }
+           }
+
+    assert mst_1918_to_1930 == %{
+             std_off: 0,
+             utc_off: -25200,
+             zone_abbr: "MST",
+             from: %{
+               wall: ~G[1918-10-27T01:00:00],
+               standard: ~G[1918-10-27T01:00:00],
+               utc: ~G[1918-10-27T08:00:00]
+             },
+             until: %{
+               wall: ~G[1930-05-04T00:00:00],
+               standard: ~G[1930-05-04T00:00:00],
+               utc: ~G[1930-05-04T07:00:00]
+             }
+           }
+  end
+
+  test "can calculate zones which start with a named rule", %{map: map} do
+    # also before GMT, so 1901-01-0T00:00:00 is actually 1900-12-31T23:00:00 UTC
+    [cet, cest | _] = calc_periods(map, "CET")
+
+    assert cet == %{
+             std_off: 0,
+             utc_off: 3600,
+             zone_abbr: "CET",
+             from: %{wall: :min, standard: :min, utc: :min},
+             until: %{
+               wall: ~G[1916-04-30T23:00:00],
+               standard: ~G[1916-04-30T23:00:00],
+               utc: ~G[1916-04-30T22:00:00]
+             }
+           }
+
+    assert cest == %{
+             std_off: 3600,
+             utc_off: 3600,
+             zone_abbr: "CEST",
+             from: %{
+               wall: ~G[1916-05-01T00:00:00],
+               standard: ~G[1916-04-30T23:00:00],
+               utc: ~G[1916-04-30T22:00:00]
+             },
+             until: %{
+               wall: ~G[1916-10-01T01:00:00],
+               standard: ~G[1916-10-01T00:00:00],
+               utc: ~G[1916-09-30T23:00:00]
+             }
+           }
+  end
+
+  test "handles rules which end with a DST transition", %{map: map} do
+    periods = calc_periods(map, "Asia/Tokyo")
+    [last_dst, last_jst] = Enum.take(periods, -2)
+
+    assert last_dst == %{
+             std_off: 3600,
+             utc_off: 32400,
+             zone_abbr: "JDT",
+             from: %{
+               wall: ~G[1951-05-06T03:00:00],
+               standard: ~G[1951-05-06T02:00:00],
+               utc: ~G[1951-05-05T17:00:00]
+             },
+             until: %{
+               wall: ~G[1951-09-08T02:00:00],
+               standard: ~G[1951-09-08T01:00:00],
+               utc: ~G[1951-09-07T16:00:00]
+             }
+           }
+
+    assert last_jst == %{
+             std_off: 0,
+             utc_off: 32400,
+             zone_abbr: "JST",
+             from: %{
+               wall: ~G[1951-09-08T01:00:00],
+               standard: ~G[1951-09-08T01:00:00],
+               utc: ~G[1951-09-07T16:00:00]
+             },
+             until: %{wall: :max, standard: :max, utc: :max}
+           }
+  end
+
+  test "handles DST transitions at the same time as zone transitions", %{map: map} do
+    periods = calc_periods(map, "America/Argentina/Buenos_Aires")
+
+    assert Enum.at(periods, 53) == %{
+             std_off: 0,
+             utc_off: -10800,
+             zone_abbr: "ART",
+             from: %{
+               utc: ~G[1993-03-07T02:00:00],
+               standard: ~G[1993-03-06T23:00:00],
+               wall: ~G[1993-03-06T23:00:00]
+             },
+             until: %{
+               utc: ~G[1999-10-03T03:00:00],
+               standard: ~G[1999-10-03T00:00:00],
+               wall: ~G[1999-10-03T00:00:00]
+             }
+           }
+
+    assert Enum.at(periods, 54) == %{
+             std_off: 3600,
+             utc_off: -14400,
+             zone_abbr: "ARST",
+             from: %{
+               utc: ~G[1999-10-03T03:00:00],
+               standard: ~G[1999-10-02T23:00:00],
+               wall: ~G[1999-10-03T00:00:00]
+             },
+             until: %{
+               utc: ~G[2000-03-03T03:00:00],
+               standard: ~G[2000-03-02T23:00:00],
+               wall: ~G[2000-03-03T00:00:00]
+             }
+           }
+  end
+
+  test "no zone periods go back in time", %{map: map} do
+    invalid_periods =
+      for zone <- map.zone_list,
+          period <- calc_periods(map, zone),
+          period.from.utc != :min,
+          period.until.utc != :max,
+          period.from.utc >= period.until.utc do
+        {zone, period}
+      end
+
+    assert invalid_periods == []
   end
 end
